+++ src/main/resources/application-dev-gcp.yml
@@ -17,8 +17,6 @@ AXSYS_CLUSTER: dev-fss
 AXSYS_NAMESPACE: org
 AXSYS_APP: axsys
 
-KAFKA_BOOTSTRAP_SERVERS: b27apvl00045.preprod.local:8443,b27apvl00046.preprod.local:8443,b27apvl00047.preprod.local:8443
-KAFKA_SCHEMA_REGISTRY_URL: https://kafka-schema-registry.nais-q.adeo.no/
 VEDTAK_FATTET_TOPIC: klage.vedtak-fattet.v1
 DVH_STATISTIKK_TOPIC: klage.kabal-statistikk.v1
